requirements.txt
asgiref==3.7.2
Django==4.2.7
django-autoslug==1.9.9
django-cities-light==3.10.0
django-tinymce==3.6.1
progressbar2==4.2.0
python-utils==3.8.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
Unidecode==1.3.7
